server: set up app to heroku
Configure the application files to a heroku deploy. A MONGODB_URI config var had to be
manually created if I wanted to proceed without inserting a credit card number.
In order to get the proper URI, I needed to go to the mlabs page, on my account,
and manually create a new app and a new user for the app.

For reference on how to set up config vars, visit the heroku support page[1]

[1]https://devcenter.heroku.com/articles/config-vars

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>