server: add token method for POST /users
This commit introduces an authentication method for users, making use of UserSchema
on the user model to create methods inside the model. The method generates a token and
adds it to the header of the REST response.

It also trims off some sensitive information from the body, by implementing
a toJSON() method inside the user model.

Ref.: Lecture 90

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>