routes: make routes to be private
In order to connect users with their todo notes, we need to configure the x-auth
token of the user with his/her notes. This commit adds such tweaks, while it adjusts
the first three tests, so they can properly test the todos tokens and associate
them with their creators.

Ref.: Lecture 99

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>