server: make remaining routes private
Make sure the remaining routes on server.js are also made private. That means
GET, DELETE and PATCH (update) routes are being modified, along with the seeding
structure, to adapt to the new implementation.
This ensures users will not be able to access todos created by other users.

Also update the server.test file to consume the changes as well.

Ref.: Lecture 100

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>