config: consume JWT_SECRET from config.json
The secret key used to sign the tokens are moved to the config.json file
so it can be hidden from production and only available to develpment and
test environments.

Heroku config JWT_SECRET key was also set but it is not included into code
as it is an environment configuration.

Ref.: Lecture 101

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>